Search bar and Show all button were added
The search bar allows users to search for workouts by the date of the workout (will work on giving the user the option to search for workouts by which type of workout it is, in the future). Also added the "Show all" button as a problem raised with the Search bar which is that once specific data is displayed, there's no way to view all the data again. This button solved that.

A BUG THAT NEEDS WORKED ON:
If the user delete a workout log for just 1 set, it deletes all the exercises. (Further troubleshooting is required).